0.5.2  Change the parameter type from int to double

+++ lib/flutter_screenutil.dart
@@ -9,8 +9,8 @@ class ScreenUtil {
   static ScreenUtil instance = new ScreenUtil();
 
   //设计稿的设备尺寸修改
-  int width;
-  int height;
+  double width;
+  double height;
   bool allowFontScaling;
 
   static MediaQueryData _mediaQueryData;
@@ -77,19 +77,19 @@ class ScreenUtil {
 
   ///根据设计稿的设备宽度适配
   ///高度也根据这个来做适配可以保证不变形
-  setWidth(int width) => width * scaleWidth;
+  setWidth(double width) => width * scaleWidth;
 
   /// 根据设计稿的设备高度适配
   /// 当发现设计稿中的一屏显示的与当前样式效果不符合时,
   /// 或者形状有差异时,高度适配建议使用此方法
   /// 高度适配主要针对想根据设计稿的一屏展示一样的效果
-  setHeight(int height) => height * scaleHeight;
+  setHeight(double height) => height * scaleHeight;
 
   ///字体大小适配方法
   ///@param fontSize 传入设计稿上字体的px ,
   ///@param allowFontScaling 控制字体是否要根据系统的“字体大小”辅助选项来进行缩放。默认值为false。
   ///@param allowFontScaling Specifies whether fonts should scale to respect Text Size accessibility settings. The default is false.
-  setSp(int fontSize) => allowFontScaling
+  setSp(double fontSize) => allowFontScaling
       ? setWidth(fontSize)
       : setWidth(fontSize) / _textScaleFactor;
 }
